imagecolor/requirements.txt
Pillow==8.3.2
matplotlib==3.4.3
tk==0.1.0
os-win==0.1.5  # or osx-photos, depending on the platform
